Store integer prices and totals and add a formatter to display it in a usual way

+++ src/main/scala/com/spikerlabs/shopping/Offers.scala
@@ -4,23 +4,23 @@ import com.spikerlabs.shopping.Offers.Offer
 
 case class CartWithOffer(cart: SomeCart, offer: Offer) extends SomeCart {
   def products: List[Product] = cart.products
-  val total: Double = cart.total - offer(cart.products)
+  val total: Int = cart.total - offer(cart.products)
 }
 
 object Offers {
-  type Offer = (List[Product]) => Double
+  type Offer = (List[Product]) => Int
   implicit class ApplyOfferToCart(cart: SomeCart) {
     def applyOffer(offer: Offer): SomeCart = CartWithOffer(cart, offer)
   }
   private def nth(n: Int, products: List[Product]) = products.drop(n - 1).grouped(n).map(_.head)
-  def buyOneGetOneFreeOnApples(products: List[Product]): Double = {
+  def buyOneGetOneFreeOnApples(products: List[Product]): Int = {
     val apples = products.flatMap {
       case apple: Apple => Some(apple)
       case _ => None
     }
     nth(2, apples).map(_.price).sum
   }
-  def threeForTwoOnOranges(products: List[Product]): Double = {
+  def threeForTwoOnOranges(products: List[Product]): Int = {
     val oranges = products.flatMap {
       case orange: Orange => Some(orange)
       case _ => None
